{"schema":"decodex/commit/1","summary":"Move macOS shell model","authority":"manual"}
